Texture aggiunte ai nemici e alla palla di fuoco,ora ci sono 20 nemici che differiscono per texture,vita,danni e alcuni per la velocità di movimento. Risolto bug nel calcolo dei danni che a volte causava il doppio dei danni

+++ project/Object.h
@@ -53,7 +53,7 @@ protected:
     float posX=0;
     float posY=0;
     std::string name;
-    int tileNumber=0;
+    float tileNumber=0;
     std::string textureFile;
 
 };
